Add temporary work-around for "run away resize" issue
Some combinations of styles and layout in a preview iframe can cause the
detected document height to always be larger than the size of the iframe,
while the code depends on the numbers matching once the iframe height is
set. This causes the iframe to grow in height infinitely; effectively
breaking the CMS for the user.

I've had a look into fixing this properly twice before and haven't found
a solution or a nice work around yet (or the exact things that cause it).
A poor man's solution is being used  as a temporary measure here so that
content can be loaded on a site that depends on this module.